cloudrun-src/requirements.txt
google-cloud-speech
websockets